trivial: rename storage_interface::state_view
In preparation for introducing DbStateView
