+++ execution/executor/src/db_bootstrapper.rs
@@ -23,7 +23,9 @@ use aptos_vm::VMExecutor;
 use executor_types::ExecutedChunk;
 use move_core_types::move_resource::MoveResource;
 use std::{collections::btree_map::BTreeMap, sync::Arc};
-use storage_interface::{state_view::VerifiedStateView, DbReaderWriter, DbWriter, TreeState};
+use storage_interface::{
+    verified_state_view::VerifiedStateView, DbReaderWriter, DbWriter, TreeState,
+};
 
 pub fn generate_waypoint<V: VMExecutor>(
     db: &DbReaderWriter,
