+++ storage/storage-interface/src/verified_state_view.rs
